Finalizing with some organizing and cleaning up of the code
