+++ src/components/Scroll.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const Scroll = (props) => {
+  return (
+    <div style={{overflowY: 'scroll', height: '500px'}}>
+        {props.children}
+    </div>
+  );
+};
+
+export default Scroll;
