+++ src/components/SearchBox.css
@@ -1,0 +1,3 @@
+input{
+    border-radius: 30px;
+}
