+++ src/components/SearchBox.js
@@ -1,9 +1,9 @@
 import React from "react";
+import './SearchBox.css'
 
 const SearchBox = ({ searchfield, searchChange }) => {
     return (
         <div className="pa2">
-
             <input
             className="pa3 ba b--green bg-lightest-blue"
             type='search'
