+++ src/containers/App.js
@@ -1,0 +1,45 @@
+import React from "react";
+import CardList from '../components/CardList';
+import SearchBox from '../components/SearchBox';
+import { Component } from 'react';
+import './App.css';
+import Scroll from '../components/Scroll';
+
+
+class App extends Component {
+    constructor() {
+        super()
+        this.state = {
+            robots: [],
+            searchfield: ''
+        }        
+    }
+
+    componentDidMount() {
+        fetch('https://jsonplaceholder.typicode.com/users')
+            .then(response=> response.json())
+            .then(users => this.setState({ robots: users}));
+    }
+
+    onSearchChange = (event) => {
+        this.setState({ searchfield: event.target.value })
+    }
+
+    render() {
+        const { robots, searchfield } = this.state;
+        const filterRobots = robots.filter(robot => {
+            return robot.name.toLowerCase().includes(searchfield.toLowerCase());
+        }) 
+        return (
+            <div className="tc">
+                <h1 className="f1">RoboFriends</h1>
+                <SearchBox searchChange={this.onSearchChange} />
+                <Scroll>
+                    <CardList robots={filterRobots}/>
+                </Scroll>                
+            </div>
+        );
+    }
+}
+
+export default App;
